Fix tests and formatting
The update to deta[async] broke several automated tests because of
differences between the old `aiodeta` api and the new official api.
The api between the async and non async versions are now more
consistent, so I removed some hacks that were required in the old test
mocks. The external API of ODetaM has not changed.

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>