Fixed passing None to `get`
Passing `None` as a key to the `get` method on either type of model
did not raise an exception directly but will either raise a weird
serialization error from deta or fail silently. This is now fixed with a
custom exception.

Note: The test for deleting an item by key had a bug where the test was trying
to save the test item's key before it was set, so it was always none. In
the previous version this produces a weird serialization error, which
caused me to find this and a larger bug around null keys. These are now
fixed with a custom exception with a nicer error message

Signed-off-by: Rick Henry <rickhenry@rickhenry.dev>